+++ Scripts/RollerBallMineValues.cs
@@ -1,0 +1,11 @@
+﻿namespace RollerBallMine.Scripts;
+
+public class RollerBallMineValues
+{
+    public float Speed;
+    public float TriggerRadius;
+    public float ExplodeTime;
+    public float ExplosionRange;
+    public int ExplosionDamage;
+    public float DetectionRange;
+}