+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### v1.0.1
+
+- Fix spawn weight
+- Balls cannot see player behind a wall or a door anymore
+- Add a sound when ball is moving
+- Update configs default values
+- Add a volume config
+- Add RandomEnemiesSize support, for big balls !
+
 ### v1.0.0
 
 - Release